Make requested changes
Following changes were made:

 - Renamed plugin to Tvibo and changed plugin matrix to reflect that

 - Added suggested error handling instead of throwing an exception

 - Removed unnecessary print statement

 - Fixed the PEP8 error
